valgrind: Add --show-leak-kinds=all and comprehensive list of suppressions.
By default valgrind suppresses many leaks.  I'm not even sure exactly
how it decides which ones to suppress, but certainly global variables
pointing to malloc’d data are suppressed, which is not useful
behaviour.

Tell valgrind to show all leaks.  It won't give an error on them.

However to do this we also need a much more comprehensive list of
suppressions so that we don't constantly show unavoidable leaks in
(eg) dlopen.  This adds separate files for each class of suppressions
in a new valgrind/ directory.

This also enables valgrinding of the Perl and Lua plugins.  Actually
this was partially enabled already, in that test-lang-plugins.c didn't
skip tests for those languages yet it was disabled in other tests.
This commit corrects that anomaly.