Add "full" plugin to complement the null, zero plugins.
It returns ENOSPC for every write request.  This is analogous to the
/dev/full device in Unix, but of course different: /dev/full is a char
device, but this is more like a thin-provisioned virtual disk which
has run out of space on the host.

You can implement this already using nbdkit-error-filter, but the
command line to do it is very long (see the man page).

This plugin has an interesting negative effect on libguestfs, because
it causes the underlying qemu to pause, causing libguestfs to hang.