null: Simplify plugin by removing some unnecessary callbacks.
There's no need for can_write since the core server will set the
NBD_FLAG_READ_ONLY flag if the server is in readonly mode (ie. ‘-r’ on
the command line).

There's no need for can_trim since this is similarly bypassed when the
server is in readonly mode.

After these changes, there's no need for the handle at all, and we can
also remove the close callback.