maint: Adjust cleaning rules
'make distcheck' calls us out for leaving files behind after
'make distclean' that were not present in the tarball.  Either
these files are expensive enough that end users should not be
required to regenerate them (so they should be distributed),
or they should be cleaned when a user asks to get back to the
pristine tarball state.

Automake suggests this hierarchy of cleaning:
mostlyclean: .o and other obvious build artifacts
clean: everything that 'make' builds without rerunning configure
distclean: everything that 'configure' builds
maintainerclean: things that requires special tooling to rebuild

It also mentions that things that are expensive to rebuild, even
if built by make and not configure, are okay in distclean - but
not in maintainerclean (this includes the TLS temporary
files, as building them consumes host entropy). Meanwhile, most
of our leaked artifacts are not terribly expensive to rebuild,
and can live in clean.  I didn't find any of the files called
out as being hard enough to regenerate to make it worth including
in the tarball, which would be the only reason to keep such files
in maintainerclean.

Signed-off-by: Eric Blake <eblake@redhat.com>