cache: Rename blk_writeback function.
The blk_writeback function was misnamed, since depending on the cache
mode and flags it could either do a writethrough or a writeback [write
to the cache only].

Since it's really the general "write a block" function, rename it
simply as blk_write.  Note there exists a blk_read function already.

There are several related changes in this commit, but it's entirely
refactoring and makes no functional difference.