cache: Implement cache-max-size and cache space reclaim.
The original plan was to have a background thread doing the reclaim.
However that cannot work given the design of filters, because a
background thread cannot access the next_ops struct which is only
available during requests.

Therefore we spread the work over the request threads.  Each blk_*
function checks whether there is work to do, and if there is will
reclaim up to two blocks from the cache (thus ensuring that we always
make forward progress, since each blk_* function can only add at most
one block to the cache).

Another large change is that the cache block size can no longer be
fixed.  We must use a block size which is at least as large as the
filesystem block size so that FALLOC_FL_PUNCH_HOLE works.  To do this,
test the filesystem block size and set blksize dynamically to
MAX (4096, filesystem block size).

This also adds a test.