common/bitmap: Add bitmap_next function and tests.
It's useful to be able to search for the next non-zero entry in a
bitmap.  This commit adds a ‘bitmap_next’ function to do that.
Because the bitmap is just a uint8_t buffer, using fast string
functions we should be able to do this quickly even if the bitmap is
sparse.  (However the actual implementation is not optimized since
that is quite complicated - see to-do comments in
common/include/nextnonzero.h).

I wasn't confident about the correctness of the code and so this
commit also adds some unit tests covering all of the bitmap code.