common: Improve pseudo-random number generation.
Currently we use non-cryptographically secure random numbers in three
places:

* the error filter, to inject errors at random,

* the partitioning plugin, to create random GUIDs, and

* the random plugin.

For these we have used either random_r or a homebrew-ish Linear
Congruential Generator.  Use of random_r is problematic on BSDs
because it doesn't exist there.  Use of the LCG was simply a bad
choice.

Replace these uses with a better quality and faster PRNG from David
Blackman and Sebastiano Vigna called ‘xoshiro256** 1.0’
(http://xoshiro.di.unimi.it/).  This is released into the public
domain (where possible) so it is compatible with the licensing of
nbdkit.

This fixes a bug in the random plugin where it could never generate
the byte 255 because I used modulo 255 instead of modulo 256
arithmetic.  Ooops.

This adds a simple sniff test that the random plugin is producing
something which at least looks like random data.

(cherry picked from commit 2ebe406f788d261522103a9a999ea158e7458354)