error: Make the random state global.
Previously the error filter held the random state in the handle, so
each connection got a separate random state, seeded when the
connection was opened.  However this makes it difficult to test the
error filter, because for example two commands like:

  qemu-io -r 'nbd:...' -c 'r 0 512'
  qemu-io -r 'nbd:...' -c 'r 0 512'

would seed the random state twice, most likely using the exact same
seed (the time in seconds).

So instead of this, use a global random state, seeded once when the
server starts.  We must hold a lock while accessing this state.

This also means we can entirely remove the error filter handle.

Note: This is not a fix.  The previous code worked fine.  This change
simply makes it easier to test.