Use linker scripts to control exported symbols.
For the server, use a linker script so that only the nbdkit_*
functions that we want to export are exported (ie. to plugins or
filters).  This prevents plugins or filters from calling arbitrary
internal server functions.

After this change only the following functions are exported by the
nbdkit binary:

$ nm -D src/nbdkit | grep ' T '
000000000000f100 T nbdkit_absolute_path
000000000000a200 T nbdkit_debug
000000000000ba90 T nbdkit_error
000000000000f3e0 T nbdkit_parse_bool
000000000000f220 T nbdkit_parse_size
000000000000f530 T nbdkit_read_password
000000000000f7e0 T nbdkit_realpath
000000000000dc30 T nbdkit_set_error
000000000000a180 T nbdkit_vdebug
000000000000ba50 T nbdkit_verror

For plugins and filters, only export the single plugin_init or
filter_init function, plus any debug flags (global variables).  For
example after this change the memory plugin only exports:

$ nm -D ./plugins/memory/.libs/nbdkit-memory-plugin.so | grep ' [BT] '
0000000000004128 B memory_debug_dir
0000000000001480 T plugin_init