tests: Fix eflags test on OpenBSD.
On OpenBSD the ‘od’ command output uses leading zeroes, eg. the
decimal number 65 is printed as "065".  Bash interprets this as octal,
and the rest of the calculations fail.  (Interestingly FreeBSD ‘od’
works the same way as Linux).

Instead print the numbers as hex and add "0x" in front.  This works
with ‘od’ across Linux, FreeBSD and OpenBSD, and the test works on all
3 platforms.