plugins, filters: Define and use NBDKIT_HANDLE_NOT_NEEDED.
A common idiom where we don't need a per-connection handle is:

  static void *
  myplugin_open (int readonly)
  {
    static int handle;
    return &handle;
  }

We need to return some non-NULL value, but we don't care what.

This commit defines a macro NBDKIT_HANDLE_NOT_NEEDED which contains a
static non-NULL pointer which can be used instead of the above.  This
is more descriptive than the previous code.