common/include: Add generic MIN and MAX macros.
The preferred implementation uses __auto_type, a GCC extension also
now supported by Clang.

Unfortunately OpenBSD ships with GCC 4.2.1 (from 2007!) which predates
this extension by quite a few years, so we have to be able to fall
back to a macro using typeof() which has been present for a long time.