tests: Test that public headers are ANSI (ISO C90) compatible.
In commit 3775916120749e935d287f35cb29ff6c586656df (and before that in
discussions with Eric) we decided that we should try to allow plugins
to be compiled with C compilers other than GCC or Clang (or with very
old / peculiar / incompatible versions of those compilers).

However until we test this we cannot guarantee that changes to the
code will not break this in future, hence this test.

Note that GCC or Clang is still required to compile nbdkit itself.