tests: cxx: Return NBDKIT_HANDLE_NOT_NEEDED from open callback.
In commit 6c4092a674b742b92d74382b3eaee53b82638776 we introduced the
concept of the open callback returning NBDKIT_HANDLE_NOT_NEEDED for
plugins which do not need a handle.  The C++ test uses this archetype
too, but in a slightly different form so that commit did not modify
this test plugin before.