build: Don't link everything, only the server, with libselinux.
The server can link to libselinux to support SELinux socket labelling.
However because we added the library to $LIBS it got linked into
everything, eg:

$ eu-readelf -d /usr/lib64/nbdkit/plugins/nbdkit-memory-plugin.so |
    grep NEEDED
  NEEDED            Shared library: [libselinux.so.1]
  NEEDED            Shared library: [libc.so.6]

Libselinux has a pkg-config file so we can use that to detect
the library and adjust server/Makefile.am so it is only linked
into the server.

This dependency is still optional.