openbsd: tests: Fix generation of file-data, offset-data.
On OpenBSD:

$ ls -l tests/file-data tests/offset-data
-rw-r--r--  1 rjones  rjones     16384 Jan 24 11:41 tests/file-data
-rw-r--r--  1 rjones  rjones  41943040 Jan 24 11:42 tests/offset-data

On Linux:

$ ls -l tests/file-data tests/offset-data
-rw-rw-r--. 1 rjones rjones     4096 Jan 24 11:41 tests/file-data
-rw-rw-r--. 1 rjones rjones 10485760 Jan 24 11:41 tests/offset-data

This is caused because we were using ‘echo -e’ to generate the files.
The OpenBSD version of /bin/echo ignores this flag and writes the
escape sequences directly into the output file.

These rules ran from the Makefile and so were using /bin/sh, not bash.
By moving the code into a bash script we can use bash, which in turn
uses the bash builtin echo command.

Note that the reason we didn't see the tests failing is because they
use libguestfs which is not present on OpenBSD, so the tests were all
skipped.

As a nice side effect of this change:

This is the last user of ‘seq’ (from coreutils) so we can remove this
dependency.  This completely removes the need for coreutils on
OpenBSD.  See also commit 738bd0d36828216ab099233fa62f22992badc2c6.