tests: Use Unix domain socket in test-partition1.sh test.
Rarely, but it did happen once, this test fails with:

  nbdkit: unable to bind to any sockets: Address already in use

This can happen if an NBD server is sitting on port 10809, or if two
test suites are running in parallel, or for other reasons.  Instead of
using a public port, make this test use a private Unix domain socket.

Fixes commit 21185fe4f29ba2b3ca4de12038023824dbea420e.