docs: Uniform documentation for plugins using magic config key.
Previously plugins which use a magic config key had a block
of text which was mostly consistent saying:

  "In nbdkit E<ge> 1.8, C<file=> may be omitted. [...]"

This commit centralises the explanation in nbdkit(1), and changes all
the man pages to (a) surround the optional key in square brackets
wherever it appears, and (b) point to the single explanation.

I've also removed the reference to nbdkit >= 1.8 because this version
is now widely available in Linux distributions, so there's less danger
of someone reading the new documentation online and trying to apply it
to an old nbdkit installed on their machine (not that anything bad
happens in this case except for an error message).