perl: Use common code for handling Perl "boolean" callbacks like can_write.
This slightly reduces the size of the generated code.

This also removes a comment I wrote when I first implemented this
plugin (which was the first "language plugin") about how language
plugins have to reproduce some internal behaviour of nbdkit.  All
language plugins have to do this, it's not news anymore.