Add ssh plugin using libssh >= 0.8.0.
This plugin allows you to access remote disk images on any server
which runs sshd.  The implementation uses sftp and works well with
OpenSSH.

libssh 0.8.0 (released Aug 2018) was the first version which
introduced the knownhosts functionality and SSH_OPTIONS_NODELAY.  The
first is quite fundamental to how the plugin works so it's worth
limiting the plugin to this minimum version.

This lacks a test.  It would be possible to test this using something
similar to what we use to test virt-p2v in libguestfs, but it's quite
tricky.