server: Minimal implementation of NBD Structured Replies.
This is about the simplest implementation of NBD Structured Replies
(SR) that we can do right now.

It accepts NBD_OPT_STRUCTURED_REPLIES when negotiated by the client,
but only sends back the simplest possible SR when required to by
NBD_CMD_READ.  The rest of the time it will send back simple replies
as before.  We do not modify the plugin API so plugins are unable to
send complex SRs.

Also we do not handle human-readable error messages yet because that
would require changes in how we handle nbdkit_error().

Also we do not understand NBD_CMD_FLAG_DF, but that seems to be OK
because (a) we don't advertize the feature and (b) we only send back a
single chunk anyway.