server: connections: Move struct connection to internal header file.
The connection struct has grown a large range of fields that are
stored per-connection but are not associated with the
server/connections.c code.  The protocol code freely accesses these
fields (because it was in the same file) but for other fields accessed
in other parts of nbdkit there were awkward accessor functions.

In this commit the structure is simply moved as a whole to
server/internal.h, thus allowing other parts of nbdkit free access.

In future it may be worth splitting this structure.