tests: rate: Fix this test to use pattern plugin instead of the memory plugin.
The rate filter test works by placing the filter above a plugin and
checking that the rate of copying is suitable throttled.

Because we were using the memory plugin - which starts off fully
sparse - there was no guarantee that the client would actually read
the plugin at all.  This is especially true when we implement Block
Status because the client now knows authoritatively that the plugin is
(initially) sparse and may skip reading it entirely.  This means that
instead of the test taking a certain minimum time to run, it runs
infinitely fast.

The simple fix here is to switch to another plugin which is known to
be non-sparse.  nbdkit-pattern-plugin or nbdkit-random-plugin would
both be reasonable choices; I used the pattern plugin.