todo: Mention possible improvement for qemu interop
One of the design aspects of structured replies was permitting a
server to reply with as much information as possible, rather than
immediately giving an error. Document one spot where we could use it
to our advantage (although qemu 3.1 won't be patched to take advantage
of it, and qemu 4.0 is being patched to not trigger the problem in the
first place, there may still be other clients that benefit).

Signed-off-by: Eric Blake <eblake@redhat.com>