Remove "All rights reserved." line from the license.
Red Hat's legal advice is that this makes NO CHANGE to the
interpretation of the license.  nbdkit remains under a very permissive
BSD-style license.  It simply removes a line in the license which was
redundant, confusing and legally meaningless.

Nevertheless we asked existing contributors for approval, and received
it from all the significant nbdkit contributors.

Acked-by: Hilko Bengen <bengen@hilluzination.de>
Acked-by: Eric Blake <eblake@redhat.com>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Nikolay Ivanets <stenavin@gmail.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>
Acked-by: Yann E. MORIN <yann.morin.1998@free.fr>
Acked-by: François Revol <revol@free.fr>
Acked-by: Nir Soffer <nirsof@gmail.com>
Acked-by: Bill Speirs <bill.speirs@gmail.com>
Acked-by: Pino Toscano <ptoscano@redhat.com>
Signed-off-by: Richard W.M. Jones <rjones@redhat.com>