tests: eflags: Use qemu-nbd to list client flags (eflags).
We previously used some hairy surgery on the oldstyle protocol to
query the nbdkit flags.  However qemu-nbd >= 4.0 can be used to query
a server’s exports (using ‘qemu-nbd --list’) and print the flags.

Since the flags now include SEND_DF unconditionally (commit c38c4c40)
I have added that to all the tests too.  It was previously not tested
because of a peculiarity of the test.