rate: Set initial bucket level correctly.
It was previously set to the capacity ratio in seconds, not the true
capacity in tokens, meaning it would always have started off nearly
empty (2 tokens, in fact).

This also affects the test.  Because the bucket is now full at the
start, the filter will burst a little initially and so the test takes
slightly less time to run (measured just over 19 seconds).  Therefore
we must adjust the test to subtract a little bit of grace time.

Fixes commit fa41958159af0c9ddaff7ae55481adebee45daeb.