Remove the xz plugin.
This was deprecated in nbdkit 1.9.2 and is scheduled for removal in
nbdkit 1.12.  This plugin has been replaced by the xz filter which is
much more capable.

Replace uses of the xz filter over files with:

  nbdkit --filter=xz file file.xz

The xz filter can also be placed on top of other plugins such as the
curl plugin.