vddk: Implement VixDiskLib_AllocateConnectParams for VDDK >= 6.7.
This avoids a warning and apparently allows us to use new features
from VDDK 6.7, although it's unclear what they might be - it seems
that QueryAllocatedBlocks works either way and that's the only new
feature we care about.

This change also stores the ConnectParams pointer in the handle and
frees it when the handle is closed.  It seems from the code sample as
if this is more correct, although documentation on this is
non-existent.