vddk: Work with VDDK 5.x.
Previously we worked with VDDK 5.x only in theory.  In practice
because we always looked for libvixDiskLib.so.6 it wouldn't have
worked.

Note even the newest VDDK 5 (version 5.5.5) was released in 2015, and
you should prefer VDDK 6 if available, not least because it enables
Flush support which is crucial for data integrity.