tests: Place sockets in /tmp.
When compiling nbdkit inside a deep directory it is easy to exceed the
maximum length of a socket path.  You will see most tests failing with
errors like:

  nbdkit: -U option: path too long (max is 107) bytes

The awkward solution to this is to place all sockets in /tmp using
`mktemp -u` to create the name.  (There is a minor and not really
exploitable race here but it's easier than creating a directory for
every test.)