ocaml: Initialize pread buffer with zeroes to avoid leaking heap memory.
In the C part of the OCaml plugin we create a ‘bytes’ [byte array] and
pass it to the OCaml pread method.  The plugin should overwrite the
array with the returned data.

However if (eg. because of a bug) the plugin does not fill the array
then whatever was in the OCaml or possibly even the C heap before the
allocation is returned to the client, possibly resulting in a leak of
sensitive data.

We can avoid this by initializing the array with zeroes.

Credit: Eric Blake for finding the bug.

Note for nbdkit 1.12: This patch is not present on the main branch.
It was posted as an early attempt to fix the OCaml bindings, but the
finally accepted approach was more complex and changes the OCaml API
(not ABI).  This patch fixes the problem as well but without changing
the API.