server: Use a thread-local pread/pwrite buffer to avoid leaking heap memory.
If the plugin .pread method did not fill the buffer with data then the
contents of the heap could be leaked back to the client.  To avoid
this create a thread-local data buffer which is initialized to zero
and expanded (with zeroes) as required.

This buffer is shared between pread and pwrite which makes the code a
little bit simpler.  Also this may improve locality by reusing the
same memory for all pread and pwrite calls in the same thread.

I have checked plugins shipped with nbdkit and none of them are
vulnerable in this way.  They all do one of three things:

 - They fully set the buffer with a single call to something like
   memcpy, memset, etc.

 - They use a loop like ‘while (count > 0)’ and correctly update count
   and the buffer pointer on each iteration.

 - For language plugins, they check that the string length returned
   from the language plugin is at least as long as the requested data,
   before memcpy-ing the data to the return buffer.

Of course I cannot check plugins which may be supplied by others.

Credit: Eric Blake for finding the bug.
(cherry picked from commit b1ce255e38251289e062bb87540f7a43c70edeae)