ocaml: Implement can_extents/extents for OCaml plugins.
Although I modified the test to add an extents method, it is currently
not called.  This is because qemu (even 4.0) does not feel the need to
issue NBD_CMD_BLOCK_STATUS calls to query the sparseness of the
underlying drive.  It's unclear if it will never do that, or if that
is simply because we are not reading the drive linearly.  It is
possible that this test will be "activated" by changes to qemu in
future.