server: extents: Set errno on error from nbdkit_add_extent.
errno is currently set to a suitable value if nbdkit_extents_new
returns an error, but this was not documented so I added that
documentation.

Set errno to a suitable value if nbdkit_add_extent returns an error,
and also document that.

Thanks: Eric Blake for spotting the problem.
(cherry picked from commit cec77449a49fd6c95b081590664ced77f3147d97)