filters: Check for mutex failures
Commit 975dab14 argued that for simple lock/unlock sequences, it was
easier to avoid the cleanup.h macros. But since that time, we added
additional sanity checking to the macros, at which point the
boilerplate of inlining that sanity checking is outweighed compared to
just using the macros in more places.

The truncate filter still uses a mutex without checks, but has bigger
issues that need to be addressed separately.

Signed-off-by: Eric Blake <eblake@redhat.com>