build: Use dlsym as sentinel function for -ldl.
When testing which “dl library” we must use for dl* symbols, autoconf
runs a test similar to:

  $ cat conftest.c
  char dlopen ();
  int main () { return dlopen (); }
  $ gcc -o conftest $CFLAGS conftest.c [try various -ldl options here]

When using ‘CFLAGS="-fsanitize=address"’ this succeeds even if no dl
libraries are used at all, since it appears that using this option
causes dlopen to be included in the final binary implicitly.  This
causes configure to set DL_LIBS='', but unfortunately dlsym is not
included implicitly and so linking fails.

Although I believe this is a bug in ASAN, as a workaround use dlsym as
the sentinel function instead.

The new output from ‘./configure CFLAGS="-fsanitize=address"’ is:

  checking for library containing dlsym... -ldl
  [...]
  checking for dladdr... yes

See also: https://bugzilla.redhat.com/show_bug.cgi?id=1702761