vddk: Do not report hole extents to be zero with single-link=true
the data in holes is actually not zeros as there might be a parent image having
other data.  Properly reporting (non-)zero hole extents allows clients to have
information about whether the data in the image is unallocated, but the read
will return zero blocks (HOLE+ZERO) or whether the data is just not in this
later of the image (only HOLE).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>