nbd: Support TCP socket
I've documented a desire to do this for a while, time to actually
follow through and support connecting as a client to a TCP server.

Note that it is desirable to support the plugin connecting to an
encrypted server over TCP, then exposing the raw data over a local
Unix socket; that aspect requires yet more work, left for another
day. But even allowing an old-style client to connect to an
unencrypted TCP server is still useful.

Signed-off-by: Eric Blake <eblake@redhat.com>