docs: Prefer 'zeroes' over 'zeros'
https://english.stackexchange.com/questions/3824/what-is-the-plural-form-of-zero
states that both 'zeroes' and 'zeros' are valid spellings;
historically, 'zeros' was the plural noun while 'zeroes' was the
third-person present tense verb (as in "when he zeroes the counter,
the digits all change to zeros"). But these days, 'zeroes' for both
forms is gaining in popularity, and is already the predominate form in
our user-facing documentation (nbdkit-plugin.pod alone using both
spellings for the noun, but with more instances of 'zeroes').  And the
NBD spec itself favors 'zeroes'.  Let's pick the form we like, and fix
the outliers for consistency.

Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit 4479590f12b2c27e6ab91d3548a3316c8adfd899)