cow: Implement .cache
Our default cow caching behavior is that if we have not yet
overwritten a portion of the image, we should pass the cache request
on to the underlying plugin (to make the upcoming reads from the
plugin possibly faster). But if we HAVE copied something locally, we
can use posix_fadvise (if available) to tell the kernel that we have
an upcoming reuse of that area of our local disk, and don't need to
bother the plugin.  This is the default because it keeps the COW image
as thin as possible.

However, another sane behavior is comparable to the 'cache' filter's
'cache_on_read' parameter: a user may be wanting to force portions of
the COW overlay to become populated, but want something more efficient
in network traffic than NBD_CMD_READ followed by NBD_CMD_WRITE of
unchanged data.  Hence, this patch also adds a 'cow_on_cache'
parameter to opt-in to the second behavior.

Signed-off-by: Eric Blake <eblake@redhat.com>