cache: Implement .cache
The whole point of the cache filter is to avoid visiting the plugin
more than once for a range of data; as such, passing cache requests
through to the plugin is wrong, and a client request for caching a
range of the file obviously means we want the data locally.  Our
filter defaults to cache_on_read=false, where we normally only cache
data that has been written but not yet flushed; but it can be presumed
that an explicit cache request should always pull data locally,
regardless of the cache_on_read setting. And even when
cache_on_read=true, we can implement caching more efficiently than
discarding the buffer of a naive pread.

Signed-off-by: Eric Blake <eblake@redhat.com>